requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
Babel==2.6.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
coverage==4.5.2
cryptography==2.4.2
cycler==0.10.0
docutils==0.14
idna==2.8
imagesize==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
more-itertools==6.0.0
mysql-connector-python==8.0.15
numpy==1.16.2
packaging==19.0
pandas==0.24.2
paramiko==2.4.2
patsy==0.5.1
pluggy==0.9.0
protobuf==3.6.1
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyNaCl==1.3.0
pyparsing==2.3.1
pytest==4.3.0
pytest-cov==2.6.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
simanneal==0.4.2
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.8.5
sphinx-autodoc-typehints==1.6.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-websupport==1.1.0
statsmodels==0.9.0
urllib3==1.24.1
